Bugfix/correct badges (#6)
* Enable Coveralls integration and update CI configurations.

Added a GitHub Actions workflow for Coveralls to track code coverage. Enhanced the infection.json5 configuration to include GitHub and Stryker settings, and updated readme badges and CI job settings for consistency.

* Update Coveralls workflow trigger and fix PHP setup

Refactor the workflow to trigger on pushes to the main branch and pull requests. Correct the PHP version setup syntax in the 'Install PHP 8.3' step.