Qa/improve workflow (#8)
* Add Release Drafter configuration and workflow

Introduce a Release Drafter to automate the release note creation process based on merged pull requests. This includes categorizing changes, autolabeling, and handling major, minor, and patch version resolutions.

* Add workflow to automate changelog updates

This workflow triggers on new releases and updates the changelog automatically. It checks out the code, updates the changelog with the latest release notes, and commits the changes back to the main branch.

* Add Dependabot auto-merge workflow

Introduce a GitHub Actions workflow to automatically merge Dependabot pull requests based on the type of version update. This configuration supports auto-merging minor, patch updates, and high-compatibility major version updates for GitHub Actions packages. Also, set up Dependabot to check for updates on a monthly schedule for GitHub Actions.

* Add GitHub workflow for updating contributors in README

This commit introduces a new workflow file to automate the updating of the collaborators and contributors sections in the README. It uses the 'contributors-readme-action' to ensure these sections are kept up to date whenever there are changes pushed to the main branch. This helps maintain an accurate list of people contributing to the project.

* Set PHP version to string format in Coveralls workflow

This change modifies the PHP version format from number to string. It ensures compatibility with the shivammathur/setup-php action, avoiding potential format issues.

* Add dependency installation step to Coveralls workflow

Added a step to install Composer dependencies before running the test coverage. This ensures that all necessary packages are available, reducing the risk of errors during the coverage generation process.